requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
discord.py>=2.3.0
schedule>=1.2.0
